backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
pymongo
google-genai
PyJWT[crypto]
pytest
pytest-cov
httpx
pillow
